Add Button to MainMenu

+++ src/tool/Tools.java
@@ -2,7 +2,7 @@ package tool;
 
 public class Tools {
 
-    public static boolean boxCollided(int x1, int y1, int w1, int h1, int x2, int y2, int w2, int h2){
+    public static boolean boxCollided(float x1, float y1, float w1, float h1, float x2, float y2, float w2, float h2){
         return x1 + w1 > x2 && x2 + w2 > x1 && y1 + h1 > y2 && y1 < y2 + h2;
     }
